Add subscription form to the top right of the site (possibly all pages).
I was only able to get the site running on my local machine well enough to test the main page. It's possible, based on the filename, that the changes will appear on all Conservancy pages. I'm not sure if that is what is wanted or not. If not, it seems like some more invasive changes would need to be made to the base template so that the form would only appear on the index page.

The form POSTs to a mailman url. Once submitted, the user is redirected to the mailman site with no obvious way (aside from the back button) to return to the main site. This can be fixed by using AJAX to POST the form and report the subscription status back to the user, but it would likely require parsing the mailman html as there doesn't appear to be any real API to mailman.

I tested this in Firefox 11.0 and Chrome 18.0.1025.113 beta.