No longer "stash" the GET object in self.
I don't think this was actually necessary ultimately.  I think the older
code, herein removed, in item_author_email() was wrong-headed in the
first place, or at the very least, was overkill.

Each item has a distinct author, according to the BlogEntry model.  So,
I think this is actually what we want.  I noticed the author filed isn't
properly going into the RSS at the moment anyway, but I'm somewhat past
caring about that, as long as the URLs now work to get author- and tag-
specific feeds.