Update Monthly Donation form to fix PayPal problem
On a monthly subscription box, PayPal silently fails to allow the user
to select any option but the first one (despite selecting another value
from the form) if you name the values the same.

It's somewhat obvious when you review the form code that PayPal gives
you that all the value="" fields were the same, and thus the incorrect
behavior is somewhat unsurprising.

I fixed this by modifying the buttons to include the amount in words.