Rework Javascript final form validation.
The problem before was that an error in the annual form would prevent
submission of the monthly form and vice-versa.  That is herein corrected
with this change, which assures that the input with id of "amount" if
the specific form (id'd with "annual" or "monthly") is the only one
checked.