Rework CSS for "Big News" & add image.
This image now is displayed with the same background and to the left of
the "Big News".  I spent extensive time researching how best to present
a larger <div> with the grey background and have the image properly
scale beside it.  Ultimately, I couldn't find a better way than this,
and this is hardly optimal.

For example, I looked into wrapping the whole thing in a div, with two
div's inside, and applying various CSS to each to get the image to
properly stay right next to the text and scale in size when resizing of
media made paragraph longer.  This generated even more problems, so I
went with the simpler solution herein, which probably isn't correct and
may well do odd things on different types of media.