Supporters list built dynamically.
Hitherto the supporters list has been committed directly to the static
sponsors/index.html file.  This was never ideal and a quick hack to
build Conservancy's supporters list at the beginning of the supporters
program.

With this change, a Django app now exists that dynamically generates the
supporters list.

The database rows must be built from Conservancy's internal Ledger file,
which will be done in a separate script.