Don't hard code style; use class and improve CSS
This changes the hard-coded style for what I'm calling the
content-with-donate-sidebar.  The advantage of not hard-coding style are
obvious, but I'm doing this now rather than later so that I can add
changes to the CSS that causes the width to extend to 100% on smaller
screen media when the donate bar disappears (the latter of which is
already implemented).