Fundraising goal lookup for template: 1st attempt
This seems to be the best approach to pass a fundraising goal record to
a template.  While the static hack that tmarble implemented probably
needs work anyway, this is probably the best way currently to interface
certain general data that we seek to place on many different pages
through the templates.

I looked into a templatetags solution, but this seemed more
straightforward and more fitting with Django principles (I think :).