Leave text w/ percentage when no progressbar
The progressbar is generated by Javascript.  As such, the text here
won't have a progressbar formatted around it when the browser doesn't
support Javascript.

However, since the Javascript does update the text, we can format the
Django template to put in place text that we wish will appear when no
Javascript is turned on, and then allow the Javascript to make it right
when it's available.

Graceful degradation: I may be its last proponent.