[PATCH 1/2] Fix unnucessary horizontal scroll bar; simplify header markup.
Because #logobutton had a fixed width, browser window narrower than 933
pixels resulted in an unnecessary horizontal scroll bar.  Fix that by
using a dynamic 100% width which results in the header image being
automatically clipped as necessary.

To avoid clipping the tree and the text though, include min-width so
that very narrow windows result in a horizontal scroll bar so the whole
name can be seen.

Furthermore, simplify the markup by removing wrapper DIV and SPAN
elements which weren’t really necessary.  As a consequence of that
change, the whole top of the page is now a link to the homepage (i.e.
‘/’).  Prior to this change, the link wasn’t clickable.