Rename Members->Projects in menu navigation et al
Due to confusion about "members" and "supporters", we want at least the
menu navigation to not use the phrase "members" since some Supporters
expect that to refer to a type of donor.

While we will not be rebranding "member projects" fully at this time, we
want to favor the term projects.

A few additional changes, in particular moving files from members/ ->
projects/ directory will be needed in next commit.