Revert back 2016 appeal to Karen's version.
Something was wrong with the version Brett committed.  It seemed to
merge text from Karen's 2016 version and the original 2015 version in
odd ways, particularly at the beginning.  This commit reverts back to
the version as Karen wanted it last night.  I will next forward-port
patches from Brett and those Karen gave this morning.