blog: Headlines clear floats.
This avoids a situation where multiple summaries have floats that stack on
top of each other vertically, which end up getting far away from the actual
text.

In order to make this work, we had to change the styling of #mainContent on
blog pages (and others with a sidebar).  Otherwise, headlines with
class="clear" would clear the sidebar as well.