Blog: add permalink on partial entries.
I noticed there is no link on the blog pages (other than on the front
page) to get the link to the blog post itself.  This adds a permalink at
the end near the tags.

I considered making the title itself a link, which is what's done on the
frontpage, but I figured this might look nicer.  Feel free to change
this if you're looking this commit message and feel I didn't DTRT
here. :)