cron: Exclude what are obviously scripts from compilation.
The impetus here is to stop the script from trying to compile and
complaining about
<www/conservancy/static/projects/policies/publish-travel-policy.py>, which
is a script written in Python 3 meant to help with repository maintenance.

Excluding files with `-` in the name seems reasonable since they can't be
modules anyway, so there's no reason to compile them.

This will start failing if we add a similar script later that doesn't have a
`-` in its name.